// pet-clinic/frontend/src/app/shared/modules/models/churndetails.ts
export interface ChurnDetails {
    churnedDate: String;
    msisdn: String;
    prodName: String;
    churnType: String;
    subChannel: String;
    churnChannel: String;
    lastValidity: String;
    amountCharged: String;
    expiryDate: String;
    totalChargedTillNow: String;
    requestDate: String;
    firstChargedDate: String;
    viaSub: String;
}